+++ plugins/api-docs/src/components/ApiExplorerTable/ApiExplorerTable.tsx
@@ -156,6 +156,7 @@ export const ApiExplorerTable = ({
 }: ExplorerTableProps) => {
   const [queryParamState, setQueryParamState] = useQueryParamState<TableState>(
     'apiTable',
+    500,
   );
 
   if (error) {